tree-linux: move all sysfs-related code to this file
This code implements Linux-specific sysfs handling for the tree API.
The Windows port provides the corresponding interfaces in tree-win.c,
so move the Linux implementation to tree-linux.c for consistency.
